subido

+++ CurraoFedericoP1PB2/TestCurraoFedericoP1PB2/curraoFedericoP1PB2/TestBiblioteca.java
@@ -97,7 +97,7 @@ public class TestBiblioteca {
 		patagonia.prestarunlibroAalumno(lucas);
 		historiasargentinas.prestarunlibroAalumno(federico);
 		
-		assertEquals(3, unlam.registroDePrestamos());
+		assertEquals((Integer)3, unlam.registroDePrestamos());
 		
 	}
